feat: mobile navbar refactoring
