+++ components/navbar/mobile/Logo.tsx
@@ -1,0 +1,10 @@
+const Logo = () => {
+  return (
+    <div>
+      <picture>
+        <img className="h-8 w-auto" src="/icon-text.svg" alt="RESF" />
+      </picture>
+    </div>
+  );
+};
+export default Logo;
