+++ components/navbar/mobile/MobileTopItems.tsx
@@ -1,26 +1,19 @@
-import { Popover } from "@headlessui/react";
-import { XMarkIcon } from "@heroicons/react/24/outline";
+import Logo from "./Logo";
 
 import type { NavItems } from "@/types/navigation/NavItem";
+import MobileMenuCloseButton from "./MobileMenuCloseButton";
 
-export interface MobileProjectItemsProps {
+export interface MobileTopItemsProps {
   items: NavItems;
 }
 
-const MobileProjectItems = ({ items }: MobileProjectItemsProps) => {
+const MobileTopItems = ({ items }: MobileTopItemsProps) => {
   return (
     <div className="px-5 pt-5 pb-6">
       <div className="flex items-center justify-between">
-        <div>
-          <picture>
-            <img className="h-8 w-auto" src="/icon-text.svg" alt="RESF" />
-          </picture>
-        </div>
+        <Logo />
         <div className="-mr-2">
-          <Popover.Button className="inline-flex items-center justify-center bg-white p-2 text-gray-400 hover:bg-gray-100 hover:text-gray-500 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-blue-500">
-            <span className="sr-only">Close menu</span>
-            <XMarkIcon className="h-6 w-6" aria-hidden="true" />
-          </Popover.Button>
+          <MobileMenuCloseButton />
         </div>
       </div>
       <div className="mt-6">
@@ -50,4 +43,4 @@ const MobileProjectItems = ({ items }: MobileProjectItemsProps) => {
   );
 };
 
-export default MobileProjectItems;
+export default MobileTopItems;
